+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- fix: SpanProtocol add setData for Swift (#1305)
+
 ## 7.2.7
 
 - fix: Remove Trace Headers below iOS 14.0 (#1309)